% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickroll-package.R
\docType{package}
\name{rickroll-package}
\alias{rickroll}
\alias{rickroll-package}
\title{rickroll: Serious rick}
\description{
It rick and rolls.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/boshek/rickroll}
  \item Report bugs at \url{https://github.com/boshek/rickroll/issues}
}

}
\author{
\strong{Maintainer}: Sam Albers \email{sam.albers@gmail.com}

}
\keyword{internal}
