% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rickroll.R
\name{rickroll}
\alias{rickroll}
\title{RickRoll}
\usage{
rickroll(return_link = TRUE)
}
\arguments{
\item{return_link}{logical, whether to return the Rick Astley link}
}
\value{
null or a URL
}
\description{
RickRoll
}
